Add "reference.html" page to generate reference audio file
The "reference.html" page features a single "Record" button that allows the user
to play the existing score with an "Organ" instrument, and save the result to
a WebM file (or OGG if WebM is not supported).

The script hardcodes the name of the score and instrument wave form to load,
but these could be parameterized later on.

The score notation is a custom-made JSON-format. File must be a JSON array.
Lines don't have any particular meaning, although it seems good practice to
use one line per measure. Notes are separated by spaces. Each note token starts
with the name of the note (e.g. "A3", or "A#3"). Duration may be specified
with "-xx", where "xx" is interpreted as 1/xxth of the duration of the round.
A final "." allows to specify dotted notes, e.g. "C4-8." is "1/8 of the round,
dotted", meaning "3/8 of the round".

Silences and plenty of other things are not yet supported. In short, that's
merely good enough for "Frère Jacques"...