Switch to Web Audio API and react to voices selection
One bug that remains: If you play with select/unselect, play/pause for some
time, the code may get lost and schedule audio at the wrong time, meaning that
audio will not be synchronized anymore. To be fixed later on...

When no singer is selected, music may take some time to resume once a player
gets selected. Normally that happens when data is not yet available, but UI
seems broken in the meantime since no audio can be heard.

Note choice of segment at which to start does not use singer metadata