[recorder] Add possibility to cancel recording and stop replay
It's easy to miss the beginning even with the 3.. 2.. 1.. countdown, or to
have a voice or sound issue in the middle of the recording. This update gives
users the ability to cancel the recording (and start again) at any time.

It also gives users the ability to stop the playback.

Note that setting `currentTime` to 0 actually does not work with the WebM files
that we have because they are not seekable. Setting the `src` attribute again
does the trick.