[player] Fix ghost voices
Voices could be incorrectly re-scheduled when removed and added again, leading
to the appearance of ghost voices after some time.

This update also removes the unnecessary call to "stop" (this probably did not
have any real consequence), and disables auto-playback when a first voice gets
added if playback was not started before.